refactor: ajuste nos tipos de valores

+++ Back-End/FruitList/Models/Fruit.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public string Description { get; set; }
-        public double A { get; set; }
-        public double B { get; set; }
+        public int A { get; set; }
+        public int B { get; set; }
     }
 }
